backend/requirements.txt
flask
cryptography
pymysql
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
google-cloud-bigquery
